+++ services/serviceaccount/CHANGELOG.md
@@ -1,3 +1,6 @@
+## v0.11.1
+- **Improvement:** Improve error handling for `CreateShortLivedAccessToken`
+
 ## v0.11.0
 - **Feature:** Add new `Algorithm` field to `CreateServiceAccountKeyPayload` model with enum values `RSA_2048` and `RSA_4096`
 
@@ -5,7 +8,7 @@
 - **Feature:** Add new enum for Key Algorithm `RSA_4096`
 
 ## v0.9.1
-  - **Dependencies:** Bump `github.com/golang-jwt/jwt/v5` from `v5.2.2` to `v5.2.3`
+- **Dependencies:** Bump `github.com/golang-jwt/jwt/v5` from `v5.2.2` to `v5.2.3`
 
 ## v0.9.0
 - Add `required:"true"` tags to model structs